- fixed issue with bsender timestamp not being casted to 64 bit early enough to send the full timestamp - preparing BPackets for being able to count the length of multiframe packets in proxy instead of my packet length hack
git-svn-id: file:///home/daniel/src/blinkenlights/legacy/svn/blinkenlights/trunk@377 fe0de55c-ccef-0310-be5c-e8635c8c2365

+++ code/blib/blib/bsender.c
@@ -438,7 +438,7 @@ b_sender_send_frame (BSender      *sender,
            /* update the timestamp */
            g_get_current_time (&tv);
            sender->packet->header.mcu_multiframe_h.timestamp 
-             = (tv.tv_sec * 1000) + (tv.tv_usec / 1000);
+             = ((guint64)tv.tv_sec * 1000) + (tv.tv_usec / 1000);
   
            /* hard-coded for 4bpp */
            sub = sender->packet->header.mcu_multiframe_h.subframe;
